+++ lua/terrortown/lang/ja.lua
@@ -1436,6 +1436,9 @@ L.searchbar_default_placeholder = "検索中..."
 --This selection of models can be extended by installing more player models.]]
 
 -- 2021-10-06
+--L.menu_server_addons_title = "Server Addons"
+--L.menu_server_addons_description = "Serverwide admin only settings for addons."
+
 --L.tooltip_finish_score_penalty_alive_teammates = "Alive teammates penalty: {score}"
 --L.finish_score_penalty_alive_teammates = "Alive teammates penalty:"
 --L.tooltip_kill_score_suicide = "Suicide: {score}"