+++ lua/terrortown/lang/zh_tw.lua
@@ -1444,6 +1444,9 @@ L.searchbar_default_placeholder = "在列表中搜索..."
 --This selection of models can be extended by installing more player models.]]
 
 -- 2021-10-06
+--L.menu_server_addons_title = "Server Addons"
+--L.menu_server_addons_description = "Serverwide admin only settings for addons."
+
 --L.tooltip_finish_score_penalty_alive_teammates = "Alive teammates penalty: {score}"
 --L.finish_score_penalty_alive_teammates = "Alive teammates penalty:"
 --L.tooltip_kill_score_suicide = "Suicide: {score}"